composekey: use a state table instead of a flat map
With a state table, we can exit out of composing as soon as we see an
invalid key (such as <Escape> or <q>), instead of waiting a constant
number of characters or expecting another press of the Compose key
(which is just confusing if you don't know whether you're composing to
begin with).

If the Compose key is a transient modifier (Alt, Shift, or Control),
trigger the Compose event on keyup instead of keydown so that the key
can still be used as the original modifier. (This is particularly
important for international layouts, where the user may want to use
the right Alt key for both AltGr and Compose.)

Remove the CapsLock key option. The IME engine still toggles the
CapsLock state even if the event is marked as handled, so with the
Compose key in this position, capitalization swings wildly.
(See https://crbug.com/339654.)

Add international layouts for the English language. (The extension
should work fine with other languages, too, but the instructions and
Options menu are not yet internationalized, and I don't want to let
manifest.json grow out of control.)