Fix up new hawaiian layout AltGr and Shift states.
Turns out we do need to keep the state of these keys in order to
properly emit the correct glyph.

This is still not ready for use as there are some issues that need to be
solved. For example, Chrome OS uses Alt + Shift + i to open the "Send
feedback" interface and unfortunately I did not see a way to disable
that. While typing in the browser Alt + e opens the "Edit" menu and
similarly I did not see a way to disable this shortcut.